Home

Yao

A General Purpose Quantum Computation Simulation Framework

Welcome to Yao, a Flexible, Extensible, Efficient Framework for Quantum Algorithm Design. Yao (幺) is the Chinese character for unitary. It is also means the first (一) in Chinese (it is the first released package from QuantumBFS).

We aim to provide a powerful tool for researchers, students to study and explore quantum computing in near term future, before quantum computer being used in large-scale.

Yao — Module.

Flexible, Extensible Framework for Quantum Algorithm Design.

Environment Variables

YaoDefaultType: set default type used in simulation.

source

Tutorial

Prepare Greenberger–Horne–Zeilinger state with Quantum Circuit

Quantum Fourier Transform

Quantum Circuit Born Machine

Manual

Interfaces

Registers

Blocks System

Block System

Cache System

Intrinsics

LuxurySparse