APIs

APIs

Yao.AbstractBlock — Type.

AbstractBlock

abstract type that all block will subtype from. N is the number of qubits.

APIs

Traits

nqubits ninput noutput isunitary ispure isreflexive ishermitian

Methods

apply! copy dispatch!

source

Yao.AbstractRegister — Type.

AbstractRegister{B, T}

abstract type that registers will subtype from. B is the batch size, T is the data type.

Required Properties

Property	Description	default

nqubits(reg)	get the total number of qubits.

nactive(reg)	get the number of active qubits.

nremain(reg)	get the number of remained qubits.	nqubits - nactive

nbatch(reg)	get the number of batch.	B

address(reg)	get the address of this register.

state(reg)	get the state of this register. It always return the matrix stored inside.

eltype(reg)	get the element type stored by this register on classical memory. (the type Julia should use to represent amplitude)	T

copy(reg)	copy this register.

similar(reg)	construct a new register with similar configuration.

Required Methods

Multiply

*(op, reg)

define how operator op act on this register. This is quite useful when there is a special approach to apply an operator on this register. (e.g a register with no batch, or a register with a MPS state, etc.)

Note

be careful, generally, operators can only be applied to a register, thus we should only overload this operation and do not overload *(reg, op).

Pack Address

pack_address!(reg, addrs)

pack addrs together to the first k-dimensions.

Example

Given a register with dimension [2, 3, 1, 5, 4], we pack [5, 4] to the first 2 dimensions. We will get [5, 4, 2, 3, 1].

Focus Address

focus!(reg, range)

merge address in range together as one dimension (the active space).

Example

Given a register with dimension (2^4)x3 and address [1, 2, 3, 4], we focus address [3, 4], will pack [3, 4] together and merge them as the active space. Then we will have a register with size 2^2x(2^2x3), and address [3, 4, 1, 2].

Initializers

Initializers are functions that provide specific quantum states, e.g zero states, random states, GHZ states and etc.

register(::Type{RT}, raw, nbatch)

an general initializer for input raw state array.

register(::Type{InitMethod}, ::Type{RT}, ::Type{T}, n, nbatch)

init register type RT with InitMethod type (e.g InitMethod{:zero}) with element type T and total number qubits n with nbatch. This will be auto-binded to some shortcuts like zero_state, rand_state, randn_state.

source

Yao.address — Function.

address(reg)->Int

get the address of this register.

source

Yao.apply! — Function.

apply!(reg, block, [signal])

apply a block to a register reg with or without a cache signal.

source

Yao.batch_normalize — Function.

batch_normalize

normalize a batch of vector.

source

Yao.batch_normalize! — Function.

batch_normalize!(matrix)

normalize a batch of vector.

source

Yao.blocks — Function.

blocks(composite_block)

get an iterator that iterate through all sub-blocks.

source

Yao.dispatch! — Method.

dispatch!(f, c, params) -> c

dispatch parameters and tweak it according to callback function f(original, parameter)->new

dispatch a vector of parameters to this composite block according to each sub-block's number of parameters.

source

Yao.iscacheable — Method.

iscaheable(server, block, level) -> Bool

whether this block is cacheable with current cache level.

source

Yao.ispure — Method.

ispure(x) -> Bool

Test whether this operator is pure.

source

Yao.isreflexive — Method.

isreflexive(x) -> Bool

Test whether this operator is reflexive.

source

Yao.isunitary — Method.

isunitary(x) -> Bool

Test whether this operator is unitary.

source

Yao.kronprod — Method.

kronprod(itr)

kronecker product all operators in the iterator.

source

Yao.log2i — Method.

log2i(x)

logrithm for integer pow of 2

source

Yao.mat — Function.

mat(block) -> Matrix

source

Yao.nactive — Function.

nactive(reg)->Int

get the number of active qubits.

source

Yao.nbatch — Function.

nbatch(reg)->Int

get the number of batch.

source

Yao.ninput — Function.

ninput(x) -> Integer

Returns the number of input qubits.

source

Yao.noutput — Function.

noutput(x) -> Integer

Returns the number of output qubits.

source

Yao.nparameters — Method.

nparameters(x) -> Integer

Returns the number of parameters of x.

source

Yao.nqubits — Function.

nqubits(x) -> Integer

Returns the total number of qubits.

source

Yao.nremain — Function.

nremain(reg)->Int

get the number of remained qubits.

source

Yao.pull — Method.

pull(server, key, pkey) -> valtype

get block's cache by (key, pkey)

source

Yao.pull — Method.

pull(server, block)

pull current block's cache from server

source

Yao.rand_state — Function.

rand_state(n, nbatch)

construct a normalized random state with uniform distributed $\theta$ and $r$ with amplitude $r\cdot e^{i\theta}$.

source

Yao.randn_state — Function.

randn_state(n, nbatch)

construct normalized a random state with normal distributed $\theta$ and $r$ with amplitude $r\cdot e^{i\theta}$.

source

Yao.register — Function.

register(::Type{RT}, raw, nbatch)

an general initializer for input raw state array.

register(::Type{InitMethod}, ::Type{RT}, ::Type{T}, n, nbatch)

init register type RT with InitMethod type (e.g InitMethod{:zero}) with element type T and total number qubits n with nbatch. This will be auto-binded to some shortcuts like zero_state, rand_state, randn_state.

source

Yao.state — Function.

state(reg)

get the state of this register. It always return the matrix stored inside.

source

Yao.statevec — Function.

statevec(reg)

get the state vector of this register. It will always return the vector form (a matrix for batched register).

source

Yao.zero_state — Function.

zero_state(n, nbatch)

construct a zero state $|00\cdots 00\rangle$.

source

Yao.ChainBlock — Type.

ChainBlock{N, T} <: CompositeBlock{N, T}

ChainBlock is a basic construct tool to create user defined blocks horizontically. It is a Vector like composite type.

source

Yao.KronBlock — Type.

KronBlock{N, T} <: CompositeBlock

composite block that combine blocks by kronecker product.

source

Yao.PhaseGate — Type.

PhiGate

Global phase gate.

source

Yao.Roller — Type.

Roller{N, M, T, BT} <: CompositeBlock{N, T}

map a block type to all lines and use a rolling method to evaluate them.

TODO

fill identity like KronBlock

source

Base.LinAlg.ishermitian — Method.

ishermitian(x) -> Bool

Test whether this operator is hermitian.

source

Base.empty! — Method.

empty!(object, signal; recursive=false)

clear this object's cache with signal, if signal < level, then do nothing.

source

Base.empty! — Method.

empty!(object; recursive=false)

force clear this object's cache

source

Base.empty! — Method.

empty!(server, key)

empty key's cache.

source

Base.empty! — Method.

empty!(type)

clear all cache in this type.

source

Base.kron — Method.

kron(blocks...) -> KronBlock
kron(iterator) -> KronBlock
kron(total, blocks...) -> KronBlock
kron(total, iterator) -> KronBlock

create a KronBlock with a list of blocks or tuple of heads and blocks.

Example

block1 = Gate(X)
block2 = Gate(Z)
block3 = Gate(Y)
KronBlock(block1, (3, block2), block3)

This will automatically generate a block list looks like

1 -- [X] --
2 ---------
3 -- [Z] --
4 -- [Y] --

source

Base.push! — Method.

push!(server, key, pkey, val[, level]) -> server

push (block, level) in hash key form to the server. update original cache with val. if input level is greater than stored level (input > stored).

source

Base.push! — Method.

push!(server, block, val, level) -> server

push val to cache server, it will be cached if level is greater than stored level. Or it will do nothing.

source

Yao.cache! — Method.

cache!(server, key, level) -> server

add a new cacheable block with cache level level by upload its key to the server.

source

Yao.cache! — Method.

cache!(server, block, level) -> server

add a new cacheable block with cache level level to the server.

source

Yao.cache_matrix — Method.

cache_matrix(block)

source

Yao.cache_type — Method.

cache_type(block) -> type

get the type that this block will use for cache.

source

Yao.define_const_gate — Method.

define_const_gate(name, ex)

define type, peroperties, matrix, etc. for this constant gate.

source

Yao.define_const_gate_new_type — Method.

define_const_gate_new_type(name, type)

define new constant matrix binding with type for this const gate. it throws UndefVarError if this constant gate is not defined.

source

Yao.define_const_gate_property — Method.

define_const_gate_property(f, property, gate, mat_ex) -> ex

define property calculated by an expression generated by function f with mat_ex for gate.

source

Yao.define_const_gate_struct — Method.

define_const_gate_struct(n, name)

define a immutable concrete type the constant gate. n is the number of qubits. name is the desired type name it will also be used as tag name, etc. It returns false if this constant gate is already defined (or this name is already used in current scope). or it returns true.

source

Yao.define_typed_const_gate — Method.

define_typed_const_gate(name, type, ex)

like define_const_gate, but this binds ex to a type specific by the user.

source

Yao.parse_block — Function.

parse_block

plugable argument transformer, overload this for different interface.

source

Yao.setlevel! — Method.

setlevel!(server, block, level)

set block's cache level

source