Blocks System

Blocks System

Blocks are the basic component of a quantum circuit in Yao.

Block System

The whole framework is consist of a block system. The whole system characterize a quantum circuit into serveral kinds of blocks. The uppermost abstract type for the whole system is AbstractBlock

Blocks

Yao.Blocks — Module.

APIs

Traits

nqubits nactive isunitary ispure isreflexive ishermitian

Methods

apply! copy dispatch!

source

Yao.Blocks.AbstractBlock — Type.

AbstractBlock

abstract type that all block will subtype from. N is the number of qubits.

source

Yao.Blocks.AbstractMeasure — Type.

AbstractMeasure <: AbstractBlock

Abstract block supertype which measurement block will inherit from.

source

Yao.Blocks.ChainBlock — Type.

ChainBlock{N, T} <: CompositeBlock{N, T}

ChainBlock is a basic construct tool to create user defined blocks horizontically. It is a Vector like composite type.

source

Yao.Blocks.CompositeBlock — Type.

CompositeBlock{N, T} <: MatrixBlock{N, T}

abstract supertype which composite blocks will inherit from.

extended APIs

blocks: get an iteratable of all blocks contained by this CompositeBlock

source

Yao.Blocks.Concentrator — Type.

Concentrator{N} <: AbstractBlock

concentrates serveral lines together in the circuit, and expose it to other blocks.

source

Yao.Blocks.ConstantGate — Type.

ConstantGate{N, T} <: PrimitiveBlock{N, T}

Abstract type for constant gates.

source

Yao.Blocks.ControlBlock — Type.

ControlBlock{BT, N, C, B, T}

BT: controlled block type, N: number of qubits, C: number of control bits, T: type of matrix.

source

Yao.Blocks.KronBlock — Type.

KronBlock{N, T} <: CompositeBlock

composite block that combine blocks by kronecker product.

source

Yao.Blocks.MatrixBlock — Type.

MatrixBlock{N, T} <: AbstractBlock

abstract type that all block with a matrix form will subtype from.

extended APIs

mat sparse full datatype

source

Yao.Blocks.PhaseGate — Type.

PhiGate

Global phase gate.

source

Yao.Blocks.PrimitiveBlock — Type.

PrimitiveBlock{N, T} <: MatrixBlock{N, T}

abstract type that all primitive block will subtype from. A primitive block is a concrete block who can not be decomposed into other blocks. All composite block can be decomposed into several primitive blocks.

NOTE: subtype for primitive block with parameter should implement hash and == method to enable key value cache.

source

Yao.Blocks.Roller — Type.

Roller{N, M, T, BT} <: CompositeBlock{N, T}

map a block type to all lines and use a rolling method to evaluate them.

TODO

fill identity like KronBlock

source

Yao.Blocks.ShiftGate — Type.

ShiftGate <: PrimitiveBlock

Phase shift gate.

source

Yao.Blocks.ConstGateTools.@const_gate — Macro.

@const_gate name[::T] = matrix

Define a new constant gate.

source

Base.LinAlg.ishermitian — Method.

ishermitian(x) -> Bool

Test whether this operator is hermitian.

source

Yao.Blocks.apply! — Function.

apply!(reg, block, [signal])

apply a block to a register reg with or without a cache signal.

source

Yao.Blocks.blocks — Function.

blocks(composite_block)

get an iterator that iterate through all sub-blocks.

source

Yao.Blocks.dispatch! — Method.

dispatch!(f, c, params) -> c

dispatch parameters and tweak it according to callback function f(original, parameter)->new

dispatch a vector of parameters to this composite block according to each sub-block's number of parameters.

source

Yao.Blocks.ispure — Method.

ispure(x) -> Bool

Test whether this operator is pure.

source

Yao.Blocks.isreflexive — Function.

isreflexive(x) -> Bool

Test whether this operator is reflexive.

source

Yao.Blocks.isunitary — Function.

isunitary(x) -> Bool

Test whether this operator is unitary.

source

Yao.Blocks.mat — Function.

mat(block) -> Matrix

source

Yao.Blocks.nparameters — Function.

nparameters(x) -> Integer

Returns the number of parameters of x.

source

Yao.Blocks.parameters — Function.

parameters(block) -> Vector

Returns a list of all parameters in block.

source

Yao.Blocks.print_block — Method.

print_block(io, block)

define the style to print this block

source