Boost

Boost

Boost is an optimization module that offers many functions for optimization.

Yao.Boost.controlled_U1 — Function.

controlled_U1(num_bit::Int, gate::AbstractMatrix, cbits::Vector{Int}, b2::Int) -> AbstractMatrix

Return general multi-controlled single qubit gate in hilbert space of num_bit qubits.

cbits specify the controling positions.

b2 is the controlled position.

source

Yao.Boost.cxgate — Method.

cxgate(::Type{MT}, num_bit::Int, b1::Ints, b2::Ints) -> PermMatrix

Single (Multiple) Controlled-X Gate on single (multiple) bits.

source

Yao.Boost.cygate — Method.

cygate(::Type{MT}, num_bit::Int, b1::Int, b2::Int) -> PermMatrix

Single Controlled-Y Gate on single bit.

source

Yao.Boost.czgate — Method.

czgate(::Type{MT}, num_bit::Int, b1::Int, b2::Int) -> Diagonal

Single Controlled-Z Gate on single bit.

source

Yao.Boost.general_controlled_gates — Method.

general_controlled_gates(num_bit::Int, projectors::Vector{Tp}, cbits::Vector{Int}, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

Return general multi-controlled gates in hilbert space of num_bit qubits,

projectors are often chosen as P0 and P1 for inverse-Control and Control at specific position.

cbits should have the same length as projectors, specifing the controling positions.

gates are a list of controlled single qubit gates.

locs should have the same length as gates, specifing the gates positions.

source

Yao.Boost.xgate — Method.

xgate(::Type{MT}, num_bit::Int, bits::Ints) -> PermMatrix

X Gate on multiple bits.

source

Yao.Boost.ygate — Method.

ygate(::Type{MT}, num_bit::Int, bits::Ints) -> PermMatrix

Y Gate on multiple bits.

source

Yao.Boost.zgate — Method.

zgate(::Type{MT}, num_bit::Int, bits::Ints) -> Diagonal

Z Gate on multiple bits.

source