Cache System

Cache System

Yao.CacheServers.AbstractCacheServer — Type.

AbstractCacheServer{K, ELT}

source

Yao.CacheServers.alloc! — Method.

alloc!(server, object, storage) -> server

alloc new storage on the server.

source

Yao.CacheServers.iscacheable — Method.

iscacheable(server, object)

check if there is available space to storage this object's value. (if this object was allocated on the server before.).

source

Yao.CacheServers.iscached — Method.

iscached(server, object, [params...])

check if this object (with params) is already cached.

source

Yao.CacheServers.pull — Method.

pull(server, object, params...) -> value

pull object storage from server.

source

Yao.CacheServers.update! — Method.

update!(storage, val) -> storage

source

Base.Distributed.clear! — Method.

clear!(server, object) -> server

clear the storage in the server of this object.

source

Base.delete! — Method.

delete!(server, object) -> server

delete this object from the server. (the storage will be deleted)

source

Base.push! — Method.

push!(server, val, object) -> server

push val to the storage of object in the server.

source