LuxurySparse

LuxurySparse

We provide more detailed optimization through a self-defined sparse library which is more efficient for operations related to quantum gates.

Yao.LuxurySparse.IMatrix — Type.

IMatrix{N, Tv}()
IMatrix{N}() where N = IMatrix{N, Int64}()
IMatrix(A::AbstractMatrix{T}) where T -> IMatrix

IMatrix matrix, with size N as label, use Int64 as its default type, both * and kron are optimized.

source

Yao.LuxurySparse.PermMatrix — Type.

PermMatrix{Tv, Ti}(perm::Vector{Ti}, vals::Vector{Tv}) where {Tv, Ti<:Integer}
PermMatrix(perm::Vector{Ti}, vals::Vector{Tv}) where {Tv, Ti}
PermMatrix(ds::AbstractMatrix)

PermMatrix represents a special kind linear operator: Permute and Multiply, which means M * v = v[perm] * val Optimizations are used to make it much faster than SparseMatrixCSC.

perm is the permutation order,

vals is the multiplication factor.

Generalized Permutation Matrix

source

Yao.LuxurySparse.pmrand — Function.

pmrand(T::Type, n::Int) -> PermMatrix

Return random PermMatrix.

source