Quantum Register

Quantum Register

Quantum Register is the abstraction of a quantum state being processed by a quantum circuit.

The Interface of Register

You can always define your own quantum register by subtyping this abstract type.

QuCircuit.AbstractRegister — Type.

AbstractRegister{B, T}

abstract type that registers will subtype from. B is the batch size, T is the data type.

Required Properties

Property	Description	default

nqubit(reg)	get the total number of qubits.

nactive(reg)	get the number of active qubits.

nremain(reg)	get the number of remained qubits.	nqubit - nactive

nbatch(reg)	get the number of batch.	B

address(reg)	get the address of this register.

state(reg)	get the state of this register. It always return the matrix stored inside.

eltype(reg)	get the element type stored by this register on classical memory. (the type Julia should use to represent amplitude)	T

copy(reg)	copy this register.

similar(reg)	construct a new register with similar configuration.

Required Methods

Multiply

*(op, reg)

define how operator op act on this register. This is quite useful when there is a special approach to apply an operator on this register. (e.g a register with no batch, or a register with a MPS state, etc.)

Note

be careful, generally, operators can only be applied to a register, thus we should only overload this operation and do not overload *(reg, op).

Pack Address

pack_address!(reg, addrs)

pack addrs together to the first k-dimensions.

Example

Given a register with dimension [2, 3, 1, 5, 4], we pack [5, 4] to the first 2 dimensions. We will get [5, 4, 2, 3, 1].

Focus Address

focus!(reg, range)

merge address in range together as one dimension (the active space).

Example

Given a register with dimension (2^4)x3 and address [1, 2, 3, 4], we focus address [3, 4], will pack [3, 4] together and merge them as the active space. Then we will have a register with size 2^2x(2^2x3), and address [3, 4, 1, 2].

Initializers

Initializers are functions that provide specific quantum states, e.g zero states, random states, GHZ states and etc.

register(::Type{RT}, raw, nbatch)

an general initializer for input raw state array.

register(::Type{InitMethod}, ::Type{RT}, ::Type{T}, n, nbatch)

init register type RT with InitMethod type (e.g InitMethod{:zero}) with element type T and total number qubits n with nbatch. This will be auto-binded to some shortcuts like zero_state, rand_state, randn_state.

source

The interface of a AbstractRegister looks like:

Properties

nqubit: number of qubits

nbatch: number of batch

nactive: number of active qubits

address: current list of line address

state: current state

eltype: eltype

copy: copy

focus!: pack several legs together

Factory Methods

QuCircuit.Register

reshaped_state: state