Array Registers

We provide ArrayReg as built in register type for simulations. It is a simple wrapper of a Julia array, e.g on CPU, we use Array by default and on CUDA devices we could use CuArray. You don't have to define your custom array type if the storage is array based.

Constructors

YaoArrayRegister.ArrayReg — Type

ArrayReg{D,T,MT<:AbstractMatrix{T}} <: AbstractArrayRegister{D}
ArrayReg{D}(raw)
ArrayReg(raw::AbstractVecOrMat; nlevel=2)
ArrayReg(r::ArrayReg)

Simulated full amplitude register type, it uses an array to represent corresponding one or a batch of quantum states. T is the numerical type for each amplitude, it is ComplexF64 by default.

Warning

ArrayReg constructor will not normalize the quantum state. If you need a normalized quantum state remember to use normalize!(register) on the register or normalize the input raw array with normalize or batched_normalize!.

We define some shortcuts to create simulated quantum states easier:

YaoArrayRegister.product_state — Function

product_state([T=ComplexF64], bit_str; nbatch=NoBatch())

Create an ArrayReg with bit string literal defined with @bit_str. See also zero_state, rand_state, uniform_state.

Examples

julia> product_state(bit"100"; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r1 = product_state(ComplexF32, bit"100"; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r2 = product_state(ComplexF32, [0, 0, 1]; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r1 ≈ r2   # because we read bit strings from right to left, vectors from left to right.
true

product_state([T=ComplexF64], total::Int, bit_config::Integer; nbatch=NoBatch(), no_transpose_storage=false)

Create an ArrayReg with bit configuration bit_config, total number of bits total. See also zero_state, rand_state, uniform_state.

Examples

julia> product_state(4, 3; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> product_state(4, 0b1001; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> product_state(ComplexF32, 4, 0b101)
ArrayReg{2, ComplexF32, Array...}
    active qubits: 4/4
    nlevel: 2

Warning

This interface will not check whether the number of required digits for the bit configuration matches the total number of bits.

YaoArrayRegister.zero_state — Function

zero_state([T=ComplexF64], n::Int; nbatch::Int=NoBatch())

Create an AbstractArrayReg with total number of bits n. See also product_state, rand_state, uniform_state and ghz_state.

Examples

julia> zero_state(4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> zero_state(ComplexF32, 4)
ArrayReg{2, ComplexF32, Array...}
    active qubits: 4/4
    nlevel: 2

julia> zero_state(ComplexF32, 4; nbatch=3)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 3

YaoArrayRegister.rand_state — Function

rand_state([T=ComplexF64], n::Int; nbatch=NoBatch(), no_transpose_storage=false)

Create a random AbstractArrayReg with total number of qudits n.

Examples

julia> rand_state(4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> rand_state(ComplexF64, 4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> rand_state(ComplexF64, 4; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

YaoArrayRegister.uniform_state — Function

uniform_state([T=ComplexF64], n; nbatch=NoBatch(), no_transpose_storage=false)

Create a uniform state: $\frac{1}{2^n} \sum_k |k⟩$. This state can also be created by applying H (Hadmard gate) on $|00⋯00⟩$ state.

Example

julia> uniform_state(4; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> uniform_state(ComplexF32, 4; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

YaoArrayRegister.oneto — Function

oneto(n::Int) -> f(register)
oneto(r::AbstractArrayReg, n::Int=nqudits(r))

Returns an register with 1:n qudits activated.

Base.repeat — Function

repeat(n, x::AbstractBlock[, locs]) -> RepeatedBlock{n}

Create a RepeatedBlock with total number of qubits n and the block to repeat on given location or on all the locations.

Example

This will create a repeat block which puts 4 X gates on each location.

julia> repeat(4, X)
nqubits: 4
repeat on (1, 2, 3, 4)
└─ X

You can also specify the location

julia> repeat(4, X, (1, 2))
nqubits: 4
repeat on (1, 2)
└─ X

But repeat won't copy the gate, thus, if it is a gate with parameter, e.g a phase(0.1), the parameter will change simultaneously.

julia> g = repeat(4, phase(0.1))
nqubits: 4
repeat on (1, 2, 3, 4)
└─ phase(0.1)

julia> g.content
phase(0.1)

julia> g.content.theta = 0.2
0.2

julia> g
nqubits: 4
repeat on (1, 2, 3, 4)
└─ phase(0.2)

Repeat over certain gates will provide speed up.

julia> reg = rand_state(20);

julia> @time apply!(reg, repeat(20, X));
  0.002252 seconds (5 allocations: 656 bytes)

julia> @time apply!(reg, chain([put(20, i=>X) for i=1:20]));
  0.049362 seconds (82.48 k allocations: 4.694 MiB, 47.11% compilation time)

repeat(x::AbstractBlock, locs)

Lazy curried version of repeat.

Properties

You can access the storage of an ArrayReg with:

YaoArrayRegister.state — Function

state(register::AbstractArrayReg) -> raw array

Returns the raw array storage of register. See also statevec.

state(ρ::DensityMatrix)

Return the raw state of density matrix ρ.

YaoArrayRegister.statevec — Function

statevec(r::ArrayReg) -> array

Return a state matrix/vector by droping the last dimension of size 1. See also state.

Warning

statevec is not type stable. It may cause performance slow down.

YaoArrayRegister.relaxedvec — Function

relaxedvec(r::AbstractArrayReg) -> AbstractArray

Return a vector representation of state, with all qudits activated. See also state, statevec.

BitBasis.hypercubic — Function

hypercubic(A::Array) -> Array

get the hypercubic representation for an array.

hypercubic(r::ArrayReg) -> AbstractArray

Return the hypercubic form (high dimensional tensor) of this register, only active qudits are considered. See also rank3.

YaoArrayRegister.rank3 — Function

rank3(r::ArrayReg)

Return the rank 3 tensor representation of state, the 3 dimensions are (activated space, remaining space, batch dimension). See also rank3.

Operations

We defined basic arithmatics for ArrayReg, besides since we do not garantee normalization for some operations on ArrayReg for simulation, normalize! and isnormalized is provided to check and normalize the simulated register.

LinearAlgebra.normalize! — Function

normalize!(r::AbstractArrayReg)

Normalize the register r in-place by its 2-norm.

YaoArrayRegister.isnormalized — Function

isnormalized(r::ArrayReg) -> Bool

Check if the register is normalized.

Specialized Instructions

We define some specialized instruction by specializing instruct! to improve the performance for simulation and dispatch them with multiple dispatch.

Implemented instruct! is listed below:

Measurement

Simulation of measurement is mainly achieved by sampling and projection.

Sample

Suppose we want to measure operational subspace, we can first get

\[p(x) = \|\langle x|\psi\rangle\|^2 = \sum\limits_{y} \|L(x, y, .)\|^2.\]

Then we sample an $a\sim p(x)$. If we just sample and don't really measure (change wave function), its over.

Projection

\[|\psi\rangle' = \sum_y L(a, y, .)/\sqrt{p(a)} |a\rangle |y\rangle\]

Good! then we can just remove the operational qubit space since x and y spaces are totally decoupled and x is known as in state a, then we get

\[|\psi\rangle'_r = \sum_y l(0, y, .) |y\rangle\]

where l = L(a:a, :, :)/sqrt(p(a)).

Others

Base.adjoint — Method

adjoint(register) -> register

Lazy adjoint for quantum registers.

Base.join — Method

join(regs...)

concatenate a list of registers regs to a larger register, each register should have the same batch size. See also clone.

BitBasis.basis — Method

basis(register) -> UnitRange

Returns an UnitRange of the all the bits in the Hilbert space of given register.

BitBasis.hypercubic — Method

hypercubic(r::ArrayReg) -> AbstractArray

Return the hypercubic form (high dimensional tensor) of this register, only active qudits are considered. See also rank3.

LinearAlgebra.normalize! — Method

normalize!(r::AbstractArrayReg)

Normalize the register r in-place by its 2-norm.

YaoAPI.focus! — Method

focus!(locs...) -> f(register) -> register

Lazy version of focus!, this returns a lambda which requires a register.

YaoAPI.partial_tr — Method

partial_tr(locs) -> f(ρ)

Curried version of partial_tr(ρ, locs).

YaoAPI.probs — Method

probs(ρ)

Returns the probability distribution from a density matrix ρ.

YaoAPI.relax! — Method

relax!(locs::Int...; to_nactive=nqudits(register)) -> f(register) -> register

Lazy version of relax!, it will be evaluated once you feed a register to its output lambda.

YaoAPI.select! — Method

select!(b::Integer) -> f(register)

Lazy version of select!. See also select.

YaoArrayRegister.arrayreg — Method

arrayreg(state; nbatch::Union{Integer,NoBatch}=NoBatch(), nlevel::Integer=2)

Create an array register, if nbatch is a integer, it will return a BatchedArrayReg.

YaoArrayRegister.arrayreg — Method

arrayreg([T=ComplexF64], bit_str; nbatch=NoBatch())

Construct an array register from bit string literal. For bit string literal please read @bit_str.

Examples

julia> arrayreg(bit"1010")
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> arrayreg(ComplexF32, bit"1010")
ArrayReg{2, ComplexF32, Array...}
    active qubits: 4/4
    nlevel: 2

YaoArrayRegister.autostatic — Method

autostatic(A[; threshold=8])

Staticize dynamic array A by a threshold.

YaoArrayRegister.batch_normalize — Function

batch_normalize

normalize a batch of vector.

YaoArrayRegister.batch_normalize! — Function

batch_normalize!(matrix)

normalize a batch of vector.

YaoArrayRegister.contiguous_shape_orders — Method

contiguous_shape_orders(shape, orders)

Merge the shape and orders if the orders are contiguous. Returns the new merged shape and order.

Example

julia> YaoArrayRegister.contiguous_shape_orders((2, 3, 4), (1, 2, 3))
([24], [1])

YaoArrayRegister.datatype — Method

datatype(register) -> Int

Returns the numerical data type used by register.

Note

datatype is not the same with eltype, since AbstractRegister family is not exactly the same with AbstractArray, it is an iterator of several registers.

YaoArrayRegister.density_fidelity — Method

density_fidelity(ρ1, ρ2)

General fidelity (including mixed states) between two density matrix for qudits.

Definition

\[F(ρ, σ)^2 = tr(ρσ) + 2 \sqrt{det(ρ)det(σ)}\]

YaoArrayRegister.exchange_sysenv — Method

exchange_sysenv(reg::AbstractArrayReg) -> AbstractRegister

Exchange system (focused qubits) and environment (remaining qubits).

YaoArrayRegister.general_controlled_gates — Method

general_controlled_gates(num_bit::Int, projectors::Vector{Tp}, cbits::Vector{Int}, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

Return general multi-controlled gates in hilbert space of num_bit qudits,

projectors are often chosen as P0 and P1 for inverse-Control and Control at specific position.

cbits should have the same length as projectors, specifing the controling positions.

gates are a list of controlled single qubit gates.

locs should have the same length as gates, specifing the gates positions.

YaoArrayRegister.ghz_state — Method

ghz_state([T=ComplexF64], n::Int; nbatch::Int=NoBatch())

Create an AbstractArrayReg with total number of bits n.

Examples

julia> ghz_state(4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

YaoArrayRegister.hilbertkron — Method

hilbertkron(num_bit::Int, gates::Vector{AbstractMatrix}, locs::Vector{Int}; nlevel=2) -> AbstractMatrix

Return general kronecher product form of gates in Hilbert space of num_bit qudits.

gates are a list of matrices.

start_locs should have the same length as gates, specifing the gates starting positions.

YaoArrayRegister.is_order_same — Method

is_order_same(locs) -> Bool

Check if the order specified by locs is the same as current order.

YaoArrayRegister.isnormalized — Method

isnormalized(r::ArrayReg) -> Bool

Check if the register is normalized.

YaoArrayRegister.linop2dense — Method

linop2dense([T=ComplexF64], linear_map!::Function, n::Int; nlevel=2) -> Matrix

Returns the dense matrix representation given linear map function.

YaoArrayRegister.matvec — Function

matvec(x::VecOrMat) -> MatOrVec

Return vector if a matrix is a column vector, else untouched.

YaoArrayRegister.most_probable — Method

most_probable(reg::AbstractArrayReg{2}, n::Int)

Find n most probable qubit configurations in a quantum register and return these configurations as a vector of BitStr instances.

YaoArrayRegister.move_ahead — Method

move_ahead(collection, orders)

Move orders to the beginning of collection.

YaoArrayRegister.mulcol! — Function

mulcol!(v::AbstractVector, i::Int, f) -> VecOrMat

multiply col i of v by f inplace.

YaoArrayRegister.mulrow! — Function

mulrow!(v::AbstractVector, i::Int, f) -> VecOrMat

multiply row i of v by f inplace.

YaoArrayRegister.mutual_information — Method

mutual_information(reg::AbstractArrayReg, part1, part2)

Compute the mutual information between locations part1 and locations part2 in a quantum state reg.

YaoArrayRegister.nbatch — Method

nbatch(register) -> Int

Returns the number of batches.

YaoArrayRegister.oneto — Method

oneto(n::Int) -> f(register)
oneto(r::AbstractArrayReg, n::Int=nqudits(r))

Returns an register with 1:n qudits activated.

YaoArrayRegister.product_state — Method

product_state([T=ComplexF64], bit_str; nbatch=NoBatch())

Create an ArrayReg with bit string literal defined with @bit_str. See also zero_state, rand_state, uniform_state.

Examples

julia> product_state(bit"100"; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r1 = product_state(ComplexF32, bit"100"; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r2 = product_state(ComplexF32, [0, 0, 1]; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2

julia> r1 ≈ r2   # because we read bit strings from right to left, vectors from left to right.
true

YaoArrayRegister.product_state — Method

product_state([T=ComplexF64], total::Int, bit_config::Integer; nbatch=NoBatch(), no_transpose_storage=false)

Create an ArrayReg with bit configuration bit_config, total number of bits total. See also zero_state, rand_state, uniform_state.

Examples

julia> product_state(4, 3; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> product_state(4, 0b1001; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> product_state(ComplexF32, 4, 0b101)
ArrayReg{2, ComplexF32, Array...}
    active qubits: 4/4
    nlevel: 2

Warning

This interface will not check whether the number of required digits for the bit configuration matches the total number of bits.

YaoArrayRegister.pure_state_fidelity — Method

pure_state_fidelity(v1::Vector, v2::Vector)

fidelity for pure states.

YaoArrayRegister.purification_fidelity — Method

purification_fidelity(m1::Matrix, m2::Matrix)

Fidelity for mixed states via purification.

Reference: http://iopscience.iop.org/article/10.1088/1367-2630/aa6a4b/meta

YaoArrayRegister.rand_state — Method

rand_state([T=ComplexF64], n::Int; nbatch=NoBatch(), no_transpose_storage=false)

Create a random AbstractArrayReg with total number of qudits n.

Examples

julia> rand_state(4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> rand_state(ComplexF64, 4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> rand_state(ComplexF64, 4; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

YaoArrayRegister.rank3 — Method

rank3(r::ArrayReg)

Return the rank 3 tensor representation of state, the 3 dimensions are (activated space, remaining space, batch dimension). See also rank3.

YaoArrayRegister.relaxedvec — Method

relaxedvec(r::AbstractArrayReg) -> AbstractArray

Return a vector representation of state, with all qudits activated. See also state, statevec.

YaoArrayRegister.sort_unitary — Method

sort_unitary(U, locations::NTuple{N, Int}) -> U

Return an sorted unitary operator according to the locations.

YaoArrayRegister.state — Method

state(register::AbstractArrayReg) -> raw array

Returns the raw array storage of register. See also statevec.

YaoArrayRegister.state — Method

state(ρ::DensityMatrix)

Return the raw state of density matrix ρ.

YaoArrayRegister.statevec — Method

statevec(r::ArrayReg) -> array

Return a state matrix/vector by droping the last dimension of size 1. See also state.

Warning

statevec is not type stable. It may cause performance slow down.

YaoArrayRegister.swapcols! — Function

swapcols!(v::VecOrMat, i::Int, j::Int[, f1, f2]) -> VecOrMat

swap col i and col j of v inplace, with f1, f2 factors applied on i and j (before swap).

YaoArrayRegister.swaprows! — Function

swaprows!(v::VecOrMat, i::Int, j::Int[, f1, f2]) -> VecOrMat

swap row i and row j of v inplace, with f1, f2 factors applied on i and j (before swap).

YaoArrayRegister.u1rows! — Function

u1rows!(state::VecOrMat, i::Int, j::Int, a, b, c, d) -> VecOrMat

apply u1 on row i and row j of state inplace.

YaoArrayRegister.uniform_state — Method

uniform_state([T=ComplexF64], n; nbatch=NoBatch(), no_transpose_storage=false)

Create a uniform state: $\frac{1}{2^n} \sum_k |k⟩$. This state can also be created by applying H (Hadmard gate) on $|00⋯00⟩$ state.

Example

julia> uniform_state(4; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

julia> uniform_state(ComplexF32, 4; nbatch=2)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 2

YaoArrayRegister.von_neumann_entropy — Method

von_neumann_entropy(reg::AbstractArrayReg, part)
von_neumann_entropy(ρ::DensityMatrix)

The entanglement entropy between part and the rest part in quantum state reg. If the input is a density matrix, it returns the entropy of a mixed state.

YaoArrayRegister.zero_state — Method

zero_state([T=ComplexF64], n::Int; nbatch::Int=NoBatch())

Create an AbstractArrayReg with total number of bits n. See also product_state, rand_state, uniform_state and ghz_state.

Examples

julia> zero_state(4)
ArrayReg{2, ComplexF64, Array...}
    active qubits: 4/4
    nlevel: 2

julia> zero_state(ComplexF32, 4)
ArrayReg{2, ComplexF32, Array...}
    active qubits: 4/4
    nlevel: 2

julia> zero_state(ComplexF32, 4; nbatch=3)
BatchedArrayReg{2, ComplexF32, Transpose...}
    active qubits: 4/4
    nlevel: 2
    nbatch: 3

YaoArrayRegister.zero_state_like — Method

zero_state_like(register, n) -> AbstractRegister

Create a register initialized to zero from an existing one.

Examples

julia> reg = rand_state(3; nbatch=2)
BatchedArrayReg{2, ComplexF64, Transpose...}
    active qubits: 3/3
    nlevel: 2
    nbatch: 2