Interfaces

Interfaces

Yao.Blocks.H — Constant.

H

The Hadamard gate acts on a single qubit. It maps the basis state $|0\rangle$ to $\frac{|0\rangle + |1\rangle}{\sqrt{2}}$ and $|1\rangle$ to $\frac{|0\rangle - |1\rangle}{\sqrt{2}}$, which means that a measurement will have equal probabilities to become 1 or 0. It is representated by the Hadamard matrix:

\[H = \frac{1}{\sqrt{2}} \begin{pmatrix} 1 & 1 \\ 1 & -1 \end{pmatrix}\]

source

Yao.Blocks.X — Constant.

X

The Pauli-X gate acts on a single qubit. It is the quantum equivalent of the NOT gate for classical computers (with respect to the standard basis $|0\rangle$, $|1\rangle$). It is represented by the Pauli X matrix:

\[X = \begin{pmatrix} 0 & 1\\ 1 & 0 \end{pmatrix}\]

source

Yao.Blocks.Y — Constant.

Y

The Pauli-Y gate acts on a single qubit. It equates to a rotation around the Y-axis of the Bloch sphere by $\pi$ radians. It maps $|0\rangle$ to $i|1\rangle$ and $|1\rangle$ to $-i|0\rangle$. It is represented by the Pauli Y matrix:

\[Y = \begin{pmatrix} 0 & -i\\ i & 0 \end{pmatrix}\]

source

Yao.Blocks.Z — Constant.

Z

The Pauli-Z gate acts on a single qubit. It equates to a rotation around the Z-axis of the Bloch sphere by $\pi$ radians. Thus, it is a special case of a phase shift gate (see shift) with $\theta = \pi$. It leaves the basis state $|0\rangle$ unchanged and maps $|1\rangle$ to $-|1\rangle$. Due to this nature, it is sometimes called phase-flip. It is represented by the Pauli Z matrix:

\[Z = \begin{pmatrix} 1 & 0\\ 0 & -1 \end{pmatrix}\]

source

Yao.Interfaces.Rx — Function.

Rx([type=Yao.DefaultType], [theta=0.0]) -> RotationGate{type, X}

Returns a rotation X gate.

source

Yao.Interfaces.Ry — Function.

Ry([type=Yao.DefaultType], [theta=0.0]) -> RotationGate{type, Y}

Returns a rotation Y gate.

source

Yao.Interfaces.Rz — Function.

Rz([type=Yao.DefaultType], [theta=0.0]) -> RotationGate{type, Z}

Returns a rotation Z gate.

source

Yao.Interfaces.chain — Function.

chain([T], n::Int) -> ChainBlock
chain([n], blocks) -> ChainBlock

Returns a ChainBlock. This factory method can be called lazily if you missed the total number of qubits.

This chains several blocks with the same size together.

source

Yao.Interfaces.concentrate — Method.

concentrate(nbit::Int, block::AbstractBlock, addrs::Vector{Int}) -> Concentrator{nbit}

concentrate blocks on serveral addrs.

source

Yao.Interfaces.control — Function.

control([total], controls, target) -> ControlBlock

Constructs a ControlBlock

source

Yao.Interfaces.phase — Function.

phase([type=Yao.DefaultType], [theta=0.0]) -> PhaseGate{:global}

Returns a global phase gate.

source

Yao.Interfaces.roll — Function.

roll([n::Int, ], blocks...) -> Roller{n}

Construct a Roller block, which is a faster than KronBlock to calculate similar small blocks tile on the whole address.

source

Yao.Interfaces.rollrepeat — Function.

rollrepeat([n::Int,] block::MatrixBlock) -> Roller{n}

Construct a Roller block, which is a faster than KronBlock to calculate similar small blocks tile on the whole address.

source

Yao.Interfaces.rot — Function.

rot([type=Yao.DefaultType], U, [theta=0.0]) -> RotationGate{type, U}

Returns an arbitrary rotation gate on U.

source

Yao.Interfaces.shift — Function.

shift([type=Yao.DefaultType], [theta=0.0]) -> PhaseGate{:shift}

Returns a phase shift gate.

source

Yao.Interfaces.swap — Function.

swap([n], [type], line1, line2) -> Swap

Returns a swap gate on line1 and line2

source

Yao.Interfaces.with! — Method.

with!(f, register)

Provide a writable context for blocks operating this register.

source

Yao.Interfaces.with — Method.

with(f, register)

Provide a copy context for blocks operating this register.

source

Yao.Blocks.XGate — Type.

XGate{T} <: ConstantGate{1, T}

The block type for Pauli-X gate. See docs for X for more information.

source

Yao.Blocks.YGate — Type.

YGate{T} <: ConstantGate{1, T}

The block type for Pauli-Y gate. See docs for Y for more information.

source

Yao.Blocks.ZGate — Type.

ZGate{T} <: ConstantGate{1, T}

The block type for Pauli-Z gate. See docs for Z for more information.

source

Base.kron — Method.

kron([total::Int, ]block0::Pair, blocks::Union{MatrixBlock, Pair}...) -> KronBlock{total}

create a KronBlock with a list of blocks or tuple of heads and blocks. If total is not provided, return a lazy constructor.

Example

kron(4, 1=>X, 3=>Z, Y)

This will automatically generate a block list looks like

1 -- [X] --
2 ---------
3 -- [Z] --
4 -- [Y] --

source

Base.repeat — Method.

repeat([n::Int,] x::MatrixBlock, [addrs]) -> RepeatedBlock{n}

Construct a RepeatedBlock, if n (the number of qubits) not supplied, using lazy evaluation. If addrs not supplied, blocks will fill the qubit space.

source