Intrinsics

Intrinsics

Yao.Intrinsics.AddressConflictError — Type.

AddressConflictError <: Exception

Address conflict error in Block Construction.

source

Yao.Intrinsics.QubitMismatchError — Type.

QubitMismatchError <: Exception

Qubit number mismatch error when applying a Block to a Register or concatenating Blocks.

source

Base.LinAlg.ishermitian — Method.

ishermitian(op) -> Bool

check if this operator is hermitian.

source

Yao.Intrinsics.basis — Method.

basis(num_bit::Int) -> UnitRange{Int}
basis(state::AbstractArray) -> UnitRange{Int}

Returns the UnitRange for basis in Hilbert Space of num_bit qubits. If an array is supplied, it will return a basis having the same size with the first diemension of array.

source

Yao.Intrinsics.batch_normalize — Function.

batch_normalize

normalize a batch of vector.

source

Yao.Intrinsics.batch_normalize! — Function.

batch_normalize!(matrix)

normalize a batch of vector.

source

Yao.Intrinsics.bdistance — Method.

bdistance(i::DInt, j::DInt) -> Int

Return number of different bits.

source

Yao.Intrinsics.bit_length — Method.

bit_length(x::Int) -> Int

Return the number of bits required to represent input integer x.

source

Yao.Intrinsics.bitarray — Method.

bitarray(v::Vector, [num_bit::Int]) -> BitArray
bitarray(v::Int, num_bit::Int) -> BitArray
bitarray(num_bit::Int) -> Function

Construct BitArray from an integer vector, if num_bit not supplied, it is 64. If an integer is supplied, it returns a function mapping a Vector/Int to bitarray.

source

Yao.Intrinsics.bmask — Function.

bmask(ibit::Int...) -> Int
bmask(bits::UnitRange{Int}) ->Int

Return an integer with specific position masked, which is offten used as a mask for binary operations.

source

Yao.Intrinsics.bsizeof — Method.

bsizeof(x) -> Int

Return the size of object, in number of bit.

source

Yao.Intrinsics.controller — Method.

controller(cbits, cvals) -> Function

Return a function that test whether a basis at cbits takes specific value cvals.

source

Yao.Intrinsics.flip — Method.

flip(index::Int, mask::Int) -> Int

Return an Integer with bits at masked position flipped.

source

Yao.Intrinsics.hilbertkron — Method.

hilbertkron(num_bit::Int, gates::Vector{AbstractMatrix}, locs::Vector{Int}) -> AbstractMatrix

Return general kronecher product form of gates in Hilbert space of num_bit qubits.

gates are a list of matrices.

start_locs should have the same length as gates, specifing the gates starting positions.

source

Yao.Intrinsics.indices_with — Method.

indices_with(num_bit::Int, poss::Vector{Int}, vals::Vector{Int}) -> Vector{Int}

Return indices with specific positions poss with value vals in a hilbert space of num_bit qubits.

source

Yao.Intrinsics.isreflexive — Method.

isreflexive(op) -> Bool

check if this operator is reflexive.

source

Yao.Intrinsics.isunitary — Method.

isunitary(op) -> Bool

check if this operator is a unitary operator.

source

Yao.Intrinsics.log2i — Method.

log2i(x::Integer) -> Integer

Return log2(x), this integer version of log2 is fast but only valid for number equal to 2^n. Ref: https://stackoverflow.com/questions/21442088

source

Yao.Intrinsics.neg — Method.

neg(index::Int, num_bit::Int) -> Int

Return an integer with all bits flipped (with total number of bit num_bit).

source

Yao.Intrinsics.onehotvec — Method.

onehotvec(::Type{T}, num_bit::Int, x::DInt) -> Vector{T}

one-hot wave vector.

source

Yao.Intrinsics.packbits — Method.

packbits(arr::AbstractArray) -> AbstractArray

pack bits to integers, usually take a BitArray as input.

source

Yao.Intrinsics.setbit — Method.

setbit(index::Int, mask::Int) -> Int

set the bit at masked position to 1.

source

Yao.Intrinsics.swapbits — Method.

swapbits(num::Int, mask12::Int) -> Int

Return an integer with bits at i and j flipped.

source

Yao.Intrinsics.takebit — Method.

takebit(index::Int, ibit::Int) -> Int

Return a bit at specific position.

source

Yao.Intrinsics.testall — Method.

testall(index::Int, mask::Int) -> Bool

Return true if all masked position of index is 1.

source

Yao.Intrinsics.testany — Method.

testany(index::Int, mask::Int) -> Bool

Return true if any masked position of index is 1.

source

Yao.Intrinsics.testval — Method.

testval(index::Int, mask::Int, onemask::Int) -> Bool

Return true if values at positions masked by mask with value 1 at positions masked by onemask and 0 otherwise.

source

Base.sort — Method.

sort(t::Tuple; lt=isless, by=identity, rev::Bool=false) -> ::Tuple

Sorts the tuple t.

source

Base.sortperm — Method.

sortperm(t::Tuple; lt=isless, by=identity, rev::Bool=false) -> ::Tuple

Computes a tuple that contains the permutation required to sort t.

source