Symbolic Computation

Symbolic Computation support for Yao

YaoSym.@ket_str — Macro

@ket_str

Create a ket register. See also @bra_str.

Example

a symbolic quantum state can be created simply by

julia> ket"110" + 2ket"111"
|110⟩ + 2|111⟩

qubits can be partially actived by focus!

julia> ket"100" + ket"111" |> focus!(1:2)
|100⟩ + |111⟩

YaoSym.@bra_str — Macro

@bra_str

Create a bra register. See also @ket_str.

Example

Similar to @ket_str literal, a symbolic quantum state can be created by

julia> bra"111" + 2bra"101"
2⟨101| + ⟨111|

julia> bra"111" * (ket"101" + ket"111")
1

Missing docstring.

Missing docstring for @vars. Check Documenter's build log for details.

YaoSym.szero_state — Function

szero_state(n; nbatch=1)

Create a symbolic zero state, same as ket"000", but allows you use an integer.