Quantum Register

Quantum Register

Quantum Register is the abstraction of a quantum state being processed by a quantum circuit.

The Interface of Register

You can always define your own quantum register by subtyping this abstract type.

QuCircuit.AbstractRegister — Type.

AbstractRegister{N, B, T}

Abstract type for quantum registers, all quantum registers contains a subtype of AbstractArray as member state.

Parameters

N is the number of qubits

B is the batch size

T eltype

source

The interface of a AbstractRegister looks like:

Properties

nqubit: number of qubits

nbatch: number of batch

nactive: number of active qubits

address: current list of line address

state: current state

eltype: eltype

copy: copy

focus!: pack several legs together

Factory Methods

QuCircuit.Register — Type.

Register{N, B, T} <: AbstractRegister{N, B, T}

default register type. This register use a builtin array to store the quantum state. The elements inside an instance of Register will be related to a certain memory address, but since it is not immutable (we need to change its shape), be careful not to change its state, though the behaviour is the same, but allocation should be avoided. Therefore, no shallow copy method is provided.

source

reshaped_state: state